Implement publishing from GitHub workflow run webhooks
This adds the "/v2/projects/<project>/versions/<version>/webhook" POST route,
which can be added to a GitHub repository as a webhook for the "workflow runs"
event.

The application will then use the information provided in the webhook to
retrieve the artifacts for the run, download them, and create a matching Build
in the database.